Fixing durable function (WIP)
