+++ samples/azure/blob-storage-databus-cleanup-function/ABSDataBus_4/DataBusBlobCleanupFunctions/host.json
@@ -1,11 +1,3 @@
 {
-  "disabled": false,
-  "version": "2.0",
-  "bindings": [
-    {
-      "name": "context",
-      "type": "orchestrationTrigger",
-      "direction": "in"
-    }
-  ]
+  "version": "2.0"
 }
