+++ samples/azure/blob-storage-databus-cleanup-function/ABSDataBus_4/DataBusBlobCleanupFunctions/local.settings.json
@@ -1,9 +1,10 @@
 {
   "IsEncrypted": false,
   "Values": {
+    "FUNCTIONS_WORKER_RUNTIME": "dotnet",
     "AzureWebJobsStorage": "UseDevelopmentStorage=true",
     "AzureWebJobsDashboard": "UseDevelopmentStorage=true",
     "DataBusStorageAccount": "UseDevelopmentStorage=true",
-    "DefaultTimeToLiveInSeconds":  "604800"
+    "DefaultTimeToLiveInSeconds": "30"
   }
 }
